projetoPROFESSOR

+++ src/main/java/br/com/api/fatec/apifatec/domain/cliente/ClienteService.java
@@ -24,21 +24,35 @@ public class ClienteService {
 	}
 
 	public void deletarCliente(Long id) {
+		Cliente cliente = encontrarClientePorId(id);
+		
+		if(cliente == null)
+			throw new IllegalArgumentException("Cliente nao existe");
+		
+		
 		clienteRepository.deleteById(id);
 	}
-
+	
 	public Cliente atualizarCliente(Long id, Cliente cliente) {
+		Cliente clienteCadastradoCliente = encontrarClientePorId(id);
 		
-		Cliente clienteCadastrado = encontrarClientePorId(id);
-		
-		if(clienteCadastrado == null) {
+		if (clienteCadastradoCliente == null)
+		{
 			return null;
-		}else {
-			clienteCadastrado.setNome(cliente.getNome());
-			clienteCadastrado.setEndereco(cliente.getEndereco());
-			clienteCadastrado.setRazaoSocial(cliente.getRazaoSocial());
-			clienteCadastrado.setEmail(cliente.getEmail());
-			return clienteRepository.save(cliente);
-		}	
+		} else {
+			clienteCadastradoCliente.setNome(cliente.getNome());
+			clienteCadastradoCliente.setRazaoSocial(cliente.getRazaoSocial());
+			clienteCadastradoCliente.setEmail(cliente.getEmail());
+			clienteCadastradoCliente.setEndereco(cliente.getEndereco());
+			return clienteRepository.save(clienteCadastradoCliente);
+		}
+	}
+
+	public List<Cliente> buscarPorNome(String nome) {
+		//Forma 1
+		return clienteRepository.findByNomeContainingIgnoreCase(nome);
+
+		//Forma 2
+		//return clienteRepository.findByNome(nome);
 	}
 }
