+++ src/main/java/br/com/api/fatec/apifatec/domain/produto/ProdutoService.java
@@ -5,8 +5,6 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import br.com.api.fatec.apifatec.domain.cliente.ClienteRepository;
-import br.com.api.fatec.apifatec.entities.Cliente;
 import br.com.api.fatec.apifatec.entities.Produto;
 
 @Service
@@ -22,29 +20,32 @@ public class ProdutoService {
 		return produtoRepository.findById(id).orElse(null);
 	}
 
-	public Produto salvarProduto(Produto produto) {
-		return produtoRepository.save(produto);
+	public Produto salvarProduto(Produto Produto) {
+		return produtoRepository.save(Produto);
 	}
 
 	public void deletarProduto(Long id) {
+		Produto Produto = encontrarProdutoPorId(id);
+		
+		if(Produto == null)
+			throw new IllegalArgumentException("Produto nao existe");
+		
+		
 		produtoRepository.deleteById(id);
 	}
-
+	
 	public Produto atualizarProduto(Long id, Produto produto) {
+		Produto ProdutoCadastrado = encontrarProdutoPorId(id);
 		
-		Produto produtoCadastrado = encontrarProdutoPorId(id);
-		
-		if(produtoCadastrado == null) {
+		if (ProdutoCadastrado == null)
+		{
 			return null;
-		}else {
-			produtoCadastrado.setNome(produto.getNome());
-			produtoCadastrado.setDescricao(produto.getDescricao());
-			produtoCadastrado.setPreco(produto.getPreco());
-			produtoCadastrado.setQtde(produto.getQtde());
-			
-			
-			return produtoRepository.save(produto);
-		}	
+		} else {
+			ProdutoCadastrado.setDescricao(produto.getDescricao());
+			ProdutoCadastrado.setPreco(produto.getPreco());
+			ProdutoCadastrado.setQuantidadeEstoque(produto.getQuantidadeEstoque());
+			ProdutoCadastrado.setAtivo(produto.getAtivo());
+			return produtoRepository.save(ProdutoCadastrado);
+		}
 	}
-
 }
